+++ NEWS.md
@@ -1,5 +1,6 @@
 # redcap.gpt (development version)
 
+* Add `fetch_redcap()` and `fetch_form()` to fetch data from the REDCap project.
 * Provided tempaltes (`.Renviron-template`, and `dev/redcap-credentials.csv`) and example usage (`analyses/connect.R`) to connect to the REDCap project.
 * Add `get_redcap_uri()`, `get_redcap_pid()`, and `get_redcap_token()` to safely read environmental variable to setup teh connection wiht the project.
 * Setup Github environment.